Create Easy
Initializing a variables 'i' and 'Length' , i used for traverse the string and Length will store length of the last world.Traverse the string until the null terminator is encountered. This loop increments i until it reaches the end of the string. To calculate lenth of the last world it finds last non space character, it moves back again to count the number of characters until it finds a space or start of string. return Length which stored the length of last word.