Create Medium
Initializing 'm' and 'n' to store the number of rows and columns in the matrix respectivelt. to represent the hight of the columns in the histogram i initialized 'height'. it iterates each row of matrix. for each row it updates height array, if current matrix is '1' , it increments the corresponding element in the heights array; otherwise, it resets it to zero.The heights array accumulates the heights of '1's encountered in each column. It employs a stack to navigate through the histogram. If the current height is smaller than the height at the top of the stack, it pops elements from the stack, calculating areas using the popped heights. The maximum area is updated whenever a larger rectangle is found. The current index is added to the stack. Once the iteration is complete, it processes any remaining elements in the stack to compute their areas.Returns the max_area, which represents the maximum area of a rectangle within the given binary matrix.