Create Hard
Iterate through the string from the end towards the beginning. Here at each iteration, it considers the substring from start of the string to current position and then it checks substring ia palindrome or not by compering with its reverse. If the substring is a palindrome, that means that the remaining part after the palindrome needs to be added in reverse order to the starting of the string to form the shortest palindrome.Return the resulting string, which includes the reverse of the remaining part after the palindrome appended to the original string.   the given string itself is a palindrome , if no palindrome prefix is found. So return the original string.